pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphbook"
version = "0.14.0-beta"
authors = [
    {name = "Richard Franklin", email = "rsamf@graphbook.ai"}
]
description = "The AI-driven data pipeline and workflow framework for data scientists and machine learning engineers."
keywords = ["ml", "workflow", "framework", "pytorch", "data science", "machine learning", "ai"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "aiohttp>=3.9.4",
    "pillow>=10.3.0",
    "watchdog>=4.0.0",
    "psutil>=6.0.0",
    "pyyaml>=6.0.2",
    "cloudpickle>=3.1.1",
    "msgpack>=1.1.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
]

[project.urls]
Homepage = "https://graphbook.ai"
Documentation = "https://docs.graphbook.ai"
Repository = "https://github.com/graphbookai/graphbook"

[project.scripts]
graphbook = "graphbook.core.cli:main"

[project.optional-dependencies]
ray = ["ray>=2.41.0"]
cv = ["torch>=2.0.0", "torchvision>=0.18.1", "opencv-python", "kornia"]
tensorboard = ["tensorboard"]
mlflow = ["mlflow"]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.8",
    "sphinxcontrib-applehelp>=1.0.8",
    "sphinxcontrib-devhelp>=1.0.6",
    "sphinxcontrib-htmlhelp>=2.0.5",
    "sphinxcontrib-jquery>=4.1",
    "sphinxcontrib-jsmath>=1.0.1",
    "sphinxcontrib-qthelp>=1.0.7",
    "sphinxcontrib-serializinghtml>=1.1.10",
    "sphinx-sitemap>=2.6.0",
    "sphinxawesome-theme>=5.2.0",
    "sphinx-design>=0.6.1",
    "selenium>=4.28.1",
    "webdriver-manager>=4.0.2",
]
peer = [
    "ray>=2.41.0",
    "torch>=2.0.0",
    "torchvision>=0.18.1",
    "s3fs>=2024.4.0",
]
